# Shorten text repr for ``DataTree``