--- a/xarray/core/datatree_render.py
+++ b/xarray/core/datatree_render.py
@@ -79,6 +79,7 @@ class RenderDataTree:
         style=None,
         childiter: type = list,
         maxlevel: int | None = None,
+        maxchildren: int | None = None,
     ):
         """
         Render tree starting at `node`.
@@ -88,6 +89,10 @@ class RenderDataTree:
                 Iterables that change the order of children  cannot be used
                 (e.g., `reversed`).
             maxlevel: Limit rendering to this depth.
+            maxchildren: Limit number of children to roughly this number. In practice,
+                for an arbitrarily large DataTree the number of children returned
+                will be (maxchildren * maxchildren - 1) / 2. The last child is also
+                included.
         :any:`RenderDataTree` is an iterator, returning a tuple with 3 items:
         `pre`
             tree prefix.
@@ -160,6 +165,14 @@ class RenderDataTree:
         root
         ├── sub0
         └── sub1
+
+        # `maxchildren` roughly limits the total number of children
+
+        >>> print(RenderDataTree(root, maxchildren=3))
+        root
+        ├── sub0
+        │   ├── sub0B
+        └── sub1
         """
         if style is None:
             style = ContStyle()
@@ -169,12 +182,17 @@ class RenderDataTree:
         self.style = style
         self.childiter = childiter
         self.maxlevel = maxlevel
+        self.maxchildren = maxchildren
 
     def __iter__(self) -> Iterator[Row]:
         return self.__next(self.node, tuple())
 
     def __next(
-        self, node: DataTree, continues: tuple[bool, ...], level: int = 0
+        self,
+        node: DataTree,
+        continues: tuple[bool, ...],
+        level: int = 0,
+        nchildren: int = 0,
     ) -> Iterator[Row]:
         yield RenderDataTree.__item(node, continues, self.style)
         children = node.children.values()
@@ -182,7 +200,18 @@ class RenderDataTree:
         if children and (self.maxlevel is None or level < self.maxlevel):
             children = self.childiter(children)
             for child, is_last in _is_last(children):
-                yield from self.__next(child, continues + (not is_last,), level=level)
+                nchildren += 1
+                if (
+                    self.maxchildren is None
+                    or nchildren < self.maxchildren
+                    or (not any(continues) and is_last)
+                ):
+                    yield from self.__next(
+                        child,
+                        continues + (not is_last,),
+                        level=level,
+                        nchildren=nchildren,
+                    )
 
     @staticmethod
     def __item(