--- a/xarray/core/formatting.py
+++ b/xarray/core/formatting.py
@@ -1137,14 +1137,24 @@ def _datatree_node_repr(node: DataTree, show_inherited: bool) -> str:
 
 def datatree_repr(dt: DataTree) -> str:
     """A printable representation of the structure of this entire tree."""
-    renderer = RenderDataTree(dt)
+    max_rows = OPTIONS["display_max_rows"]
+
+    renderer = RenderDataTree(dt, maxchildren=max_rows)
 
     name_info = "" if dt.name is None else f" {dt.name!r}"
     header = f"<xarray.DataTree{name_info}>"
 
     lines = [header]
     show_inherited = True
-    for pre, fill, node in renderer:
+
+    rendered_items = list(renderer)
+    for i, (pre, fill, node) in enumerate(rendered_items):
+        if len(rendered_items) > max_rows:
+            if i == max_rows:
+                lines.append("...")
+            if i >= max_rows and i != (len(rendered_items) - 1):
+                continue
+
         node_repr = _datatree_node_repr(node, show_inherited=show_inherited)
         show_inherited = False  # only show inherited coords on the root