## Demo

```python
import numpy as np
import xarray as xr

number_of_files = 700
number_of_groups = 5
number_of_variables= 10

datasets = {}
for f in range(number_of_files):
    for g in range(number_of_groups):
        # Create random data
        time = np.linspace(0, 50 + f, 1 + 1000 * g)
        y = f * time + g

        # Create dataset:
        ds = xr.Dataset(
            data_vars={
                f"temperature_{g}{i}": ("time", y)
                for i in range(number_of_variables // number_of_groups)
            },
            coords={"time": ("time", time)},
        ).chunk()

        # Prepare for xr.DataTree:
        name = f"file_{f}/group_{g}"
        datasets[name] = ds


dt = xr.DataTree.from_dict(datasets)

%timeit dt._repr_html_()
# 62.6 ms ± 1.18 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit dt.__repr__()
# 3.23 ms ± 66.5 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Here is what the repr actually looks like:
```python
In [2]: dt
Out[2]: 
<xarray.DataTree>
Group: /
├── Group: /file_0
│   ├── Group: /file_0/group_0
│   │       Dimensions:         (time: 1)
│   │       Coordinates:
│   │         * time            (time) float64 8B 0.0
│   │       Data variables:
│   │           temperature_00  (time) float64 8B dask.array<chunksize=(1,), meta=np.ndarray>
│   │           temperature_01  (time) float64 8B dask.array<chunksize=(1,), meta=np.ndarray>
│   ├── Group: /file_0/group_1
│   │       Dimensions:         (time: 1001)
│   │       Coordinates:
│   │         * time            (time) float64 8kB 0.0 0.05 0.1 0.15 ... 49.9 49.95 50.0
│   │       Data variables:
│   │           temperature_10  (time) float64 8kB dask.array<chunksize=(1001,), meta=np.ndarray>
│   │           temperature_11  (time) float64 8kB dask.array<chunksize=(1001,), meta=np.ndarray>
│   ├── Group: /file_0/group_2
│   │       Dimensions:         (time: 2001)
│   │       Coordinates:
│   │         * time            (time) float64 16kB 0.0 0.025 0.05 ... 49.95 49.98 50.0
│   │       Data variables:
│   │           temperature_20  (time) float64 16kB dask.array<chunksize=(2001,), meta=np.ndarray>
│   │           temperature_21  (time) float64 16kB dask.array<chunksize=(2001,), meta=np.ndarray>
│   ├── Group: /file_0/group_3
│   │       Dimensions:         (time: 3001)
│   │       Coordinates:
│   │         * time            (time) float64 24kB 0.0 0.01667 0.03333 ... 49.97 49.98 50.0
│   │       Data variables:
│   │           temperature_30  (time) float64 24kB dask.array<chunksize=(3001,), meta=np.ndarray>
│   │           temperature_31  (time) float64 24kB dask.array<chunksize=(3001,), meta=np.ndarray>
│   └── Group: /file_0/group_4
│           Dimensions:         (time: 4001)
│           Coordinates:
│             * time            (time) float64 32kB 0.0 0.0125 0.025 ... 49.98 49.99 50.0
│           Data variables:
│               temperature_40  (time) float64 32kB dask.array<chunksize=(4001,), meta=np.ndarray>
│               temperature_41  (time) float64 32kB dask.array<chunksize=(4001,), meta=np.ndarray>
├── Group: /file_1
│   ├── Group: /file_1/group_0
│   │       Dimensions:         (time: 1)
│   │       Coordinates:
│   │         * time            (time) float64 8B 0.0
│   │       Data variables:
│   │           temperature_00  (time) float64 8B dask.array<chunksize=(1,), meta=np.ndarray>
│   │           temperature_01  (time) float64 8B dask.array<chunksize=(1,), meta=np.ndarray>
│   ├── Group: /file_1/group_1
│   │       Dimensions:         (time: 1001)
│   │       Coordinates:
│   │         * time            (time) float64 8kB 0.0 0.051 0.102 0.153 ... 50.9 50.95 51.0
│   │       Data variables:
│   │           temperature_10  (time) float64 8kB dask.array<chunksize=(1001,), meta=np.ndarray>
│   │           temperature_11  (time) float64 8kB dask.array<chunksize=(1001,), meta=np.ndarray>
│   ├── Group: /file_1/group_2
│   │       Dimensions:         (time: 2001)
│   │       Coordinates:
│   │         * time            (time) float64 16kB 0.0 0.0255 0.051 ... 50.95 50.97 51.0
│   │       Data variables:
│   │           temperature_20  (time) float64 16kB dask.array<chunksize=(2001,), meta=np.ndarray>
│   │           temperature_21  (time) float64 16kB dask.array<chunksize=(2001,), meta=np.ndarray>
│   ├── Group: /file_1/group_3
│   │       Dimensions:         (time: 3001)
│   │       Coordinates:
│   │         * time            (time) float64 24kB 0.0 0.017 0.034 ... 50.97 50.98 51.0
│   │       Data variables:
│   │           temperature_30  (time) float64 24kB dask.array<chunksize=(3001,), meta=np.ndarray>
│   │           temperature_31  (time) float64 24kB dask.array<chunksize=(3001,), meta=np.ndarray>
...
    └── Group: /file_699/group_4
            Dimensions:         (time: 4001)
            Coordinates:
              * time            (time) float64 32kB 0.0 0.1872 0.3745 ... 748.6 748.8 749.0
            Data variables:
                temperature_40  (time) float64 32kB dask.array<chunksize=(4001,), meta=np.ndarray>
                temperature_41  (time) float64 32kB dask.array<chunksize=(4001,), meta=np.ndarray>
```

Here is the html repr:

![image](https://github.com/user-attachments/assets/aae67e38-6414-41c0-9ac2-b9573b1221ad)

## Notes

This logic is configurable via `xr.set_options(display_max_rows=)` I decide to reuse that option rather than making a new one since it felt functionally kind of similar to limiting the number of variables or coords shown on a dataset (which is what `display_max_rows` is otherwise used for). I can imagine making a new `display_max_children` setting though and I'd likely toggle the default 12 down to 6 or something. 

I wasn't quite sure what the right approach was so I tried two different ideas. It might make sense to consolidate on one approach though

In the text repr:

1. The max is applied to the total number of nodes, so the overall size of the output repr is fixed at n nodes. 
2. The un-repred nodes are replaced with a "..."
3. The final node is always displayed

In the html repr:

1. The max is applied to the number of children, so any given node can have up to n children
2. There is a note about how not all nodes are displayed

## Checklist


- [x] Closes [#10052](https://github.com/pydata/xarray/issues/10052)
- [x] Tests added
- [ ] User visible changes (including notable bug fixes) are documented in `whats-new.rst`