@TomNicholas 2025-03-18
Thank you for working on this @jsignell !

> %timeit dt._repr_html_()
> # 62.6 ms ± 1.18 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
>
> %timeit dt.__repr__()
> # 3.23 ms ± 66.5 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

This seems good! But how much better is it than before?

> There is a note about how not all nodes are displayed

This seems a little ugly from a UI perspective.