@benbovy 2025-03-19
Looks nice @jsignell! Some thoughts:

>  I decide to reuse that option rather than making a new one since it felt functionally kind of similar to limiting the number of variables or coords shown on a dataset (which is what display_max_rows is otherwise used for). I can imagine making a new display_max_children setting though and I'd likely toggle the default 12 down to 6 or something.

Same feelings and no strong opinion on this. I'm slightly leaning towards a separate option, though, in order to avoid constantly toggling the value of `display_max_rows` when working with both dataset and datatree objects.

> The final node is always displayed

Could we instead compute the location of the "..." placeholder so that it is more in the "middle", likely showing more than one trailing node? It is probably better from a UX perspective (and more consistent with other reprs such as Dataset and `pandas.DataFrame`)

> It might make sense to consolidate on one approach though

I agree. Maybe adding "..." placeholder rows in the html repr too and reuse the same logic for computing their location?

> There is a note about how not all nodes are displayed

Would it be easy to customize the number of items in the section title rather than adding a full note? E.g., something like:

```
 ▶ Groups:  (700 / 12 displayed)
```

or just

```
 ▶ Groups:  (12/700)
```