@Illviljan 2025-04-18
The benchmark compares the PR to Main on the same machine. And it looks good!  
```
| Change   | Before [38169010]    | After [a06b2c1b]    |   Ratio | Benchmark (Parameter)            |
|----------|----------------------|---------------------|---------|----------------------------------|
| -        | 54.9±0.5ms           | 4.40±0.03ms         |    0.08 | repr.ReprDataTree.time_repr      |
| -        | 186±2ms              | 13.8±0.1ms          |    0.07 | repr.ReprDataTree.time_repr_html |
```

Note that the ASV test is a little smaller than the intro example. 
https://github.com/pydata/xarray/blob/38169010846e772a085158c356ef2ed6b8f2086a/asv_bench/benchmarks/repr.py#L28-L38